feat(nav): move namespace filter into the cluster scope zone + make it view-aware
Relocate the namespace switcher from the right utility group to a left
"scope zone" beside the cluster switcher, rendered as a lighter, trailing,
view-aware chip — namespace and cluster together answer "what am I looking
at", so they belong adjacent (kubectl --context X -n Y), while the lighter
weight signals a reversible view filter vs. a destructive context switch.

View-aware: the chip is disabled (with an explanation) on cluster-scoped
surfaces where it was previously a dead control — the Cost view, GitOps
detail trees, and cluster-scoped resource kinds (Nodes, PVs, ClusterRoles).
The pick is preserved so it re-applies on return to a namespaced view.

Extract the trigger+dropdown into a shared presentational NamespacePicker in
@skyhook-io/k8s-ui (mirrors the ClusterSwitcher pattern — pure UI, data via
props) and re-export it from @skyhook-io/radar-app, so Radar Hub can render a
namespace filter identical to OSS driving its own per-cluster scope. OSS's
NamespaceSwitcher becomes a thin data container over the primitive.
