+++ packages/k8s-ui/src/components/namespace-switcher/index.ts
@@ -1,0 +1,6 @@
+export { NamespacePicker } from './NamespacePicker'
+export type {
+  NamespacePickerProps,
+  NamespacePickerHandle,
+  NamespaceScopeView,
+} from './NamespacePicker'
